refactor(phi_moe): specialize family implementation
Consolidate Phi-MoE graph construction, decoder runtime, tensor-parallel building, and weight loading under its owned canonical packages. Preserve the stable debug adapter and lazy plugin API while removing unreachable family-local decoder copies.
