+++ tests/e2e/models/phi_moe/test_phi_moe_builder_tp.py
@@ -17,7 +17,7 @@ pytest.importorskip("tensorrt", reason="TensorRT is required for family builder 
 try:
     phi_moe_module = importlib.import_module(
         "tensorrt_model_connect.families.phi_moe.plugin")
-    from tensorrt_model_connect.families.phi_moe import tp_builder
+    from tensorrt_model_connect.families.phi_moe.model import parallel as tp_builder
     from tensorrt_model_connect.parallel_config import ParallelConfig
 except (ImportError, ModuleNotFoundError):
     pytest.skip("tensorrt_model_connect requires tensorrt", allow_module_level=True)
